refactor: rename SSO files
